extract Soup filter cache adapter
