+++ .config/hakari.toml
@@ -83,6 +83,7 @@ workspace-members = [
     "macro_uuid",
     "non_empty",
     "cowlike",
+    "soup-filter-cache-adapter",
 ]
 
 # Xtask extension (the hakari engine ignores this table; `cargo x deps` in
@@ -93,4 +94,4 @@ workspace-members = [
 # via cache-wasm; cache-sqlite ships in the Tauri app, not services). They
 # must stay wasm-safe and out of the deploy dependency layers — the glob
 # keeps future client crates from silently gaining the workspace-hack dep.
-traversal-exclude-globs = ["crates/client/*"]
+traversal-exclude-globs = ["crates/client/*", "crates/soup_filter_cache_adapter*"]
