+++ crates/soup_filter_cache_adapter/src/test.rs
@@ -1,0 +1,77 @@
+use super::*;
+
+#[test]
+fn optimistic_soup_payloads_compile_to_durable_projection_layers() {
+    let mutations = optimistic_projection_mutations(
+        &serde_json::json!({
+            "update": {
+                "__typename": "GraphqlSoupDocument",
+                "id": "00000000-0000-0000-0000-000000000001",
+                "projectId": "00000000-0000-0000-0000-000000000002"
+            }
+        }),
+        123,
+    );
+    assert_eq!(mutations.len(), 1);
+    let OptimisticProjectionMutation::Patch {
+        exact,
+        integers,
+        sorts,
+        ..
+    } = &mutations[0]
+    else {
+        panic!("partial optimistic Soup entity should compile to a patch");
+    };
+    assert!(
+        exact
+            .iter()
+            .any(|patch| patch.attribute == vocabulary::project_id())
+    );
+    assert!(integers.iter().any(|patch| {
+        patch.attribute == vocabulary::updated_at() && patch.values == vec![123_000]
+    }));
+    assert!(
+        sorts
+            .iter()
+            .any(|fact| { fact.attribute == vocabulary::updated_at() && fact.value == 123_000 })
+    );
+
+    let complete = optimistic_projection_mutations(
+        &serde_json::json!({
+            "create": {
+                "__typename": "GraphqlSoupProject",
+                "id": "00000000-0000-0000-0000-000000000003",
+                "ownerId": "user-1",
+                "parentId": null,
+                "createdAt": "2025-01-01T00:00:00.000001Z"
+            }
+        }),
+        456,
+    );
+    assert!(matches!(
+        complete.as_slice(),
+        [OptimisticProjectionMutation::Replace(_)]
+    ));
+
+    let deletion = optimistic_projection_mutations(
+        &serde_json::json!({
+            "delete": {
+                "__typename": "GraphqlCacheDeletion",
+                "graphqlTypeName": "GraphqlSoupChat",
+                "entityId": "00000000-0000-0000-0000-000000000004"
+            }
+        }),
+        789,
+    );
+    assert!(matches!(
+        deletion.as_slice(),
+        [OptimisticProjectionMutation::Delete { .. }]
+    ));
+}
+
+#[test]
+fn invalid_sort_direction_is_rejected_at_the_soup_boundary() {
+    let error = compile_filter_request(serde_json::json!({}), "UPDATED_AT", "SIDEWAYS", 10)
+        .expect_err("invalid direction must not reach the generic cache");
+    assert_eq!(error.to_string(), "invalid entity-filter sort direction");
+}
